fixed the tests renaming
